resources: update MAME files and add a generator
Added a script to generate the MAME related resources (devices/bioses/names) from a series of DAT files.
`mameres.py` ingests several XML based DAT files and produces the 3 MAME resource files. The operation merges the info from all given files.
For `mamenames.xml`, the `realname` precendence is based on the 1st occurence of the romset in the DAT files, so the order of the files is important.

The resources added in this commit are based on (in this order):
 * MAME 240 (Jan-2022) (games, bioses and devices)
 * FinalBurn Neo (11 Feb 2022) (games and bioses)
 * Mame2003 Libretro (11 Feb 2022) (games and bioses)

Note: the devices file was generated from the 'MAME 0.240.dat' file, since it contains both MESS and Arcade systems.