input: remove useless XML node in the temporary output
The `<inputAction>` node is useless for the temporary input configuration, so remove it.

Normally, because of the previous `reset` call of the parent document, it would be empty, but I've encountered situations when the memory is not cleared and there's (random binary) garbage added to the temporary configuration file. This results in a broken XML file and the auto-configuration scripts in RetroPie will all fail.