Gamelist bugfix: don't show un-registered extensions
Updated the parsing of an existing gamelist to exclude entries without an extension present in `es_systems.cfg`.
This can happen when the list of extensions is updated and some extensions are removed, but the `gamelist.xml` entries are not removed. EmulationStation would still consider the excluded extension's entries as valid gamelist entries, without filtering them out.